requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.5.2
argcomplete==1.10.3
asgiref==3.8.1
astor==0.8.1
beautifulsoup4==4.8.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
colorama==0.4.6
compressed_rtf==1.0.6
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
distro==1.9.0
Django==5.1.6
docx2txt==0.8
ebcdic==1.1.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
extract-msg==0.28.7
fire==0.7.0
fonttools==4.55.8
git-filter-repo==2.47.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
IMAPClient==2.1.0
imgaug==0.4.0
jiter==0.8.2
joblib==1.4.2
jsonpath-python==1.0.6
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.0
matplotlib==3.10.0
mistralai==1.5.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
olefile==0.47
openai==1.61.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20191110
pillow==11.1.0
protobuf==3.20.2
pyclipper==1.3.0.post6
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
requests==2.32.3
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
shapely==2.0.7
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
sqlparse==0.5.3
termcolor==2.5.0
textract==1.6.5
threadpoolctl==3.5.0
tifffile==2025.1.10
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
xlrd==1.2.0
XlsxWriter==3.2.2
